requirements.txt
django
gunicorn
django-heroku
py3DNS
biopython
pydna